requirements.txt
2captcha-python==1.0.3
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
futu-api==5.1.1320
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
protobuf==3.5.1
psutil==5.8.0
pycryptodome==3.10.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
Werkzeug==1.0.1
